floating point calculations in c acheived

+++ c_float/Policies/PureEndowment.c
@@ -12,52 +12,52 @@
 #include "../RK_Library.h"
 #include "PureEndowment.h"
 
-static double *matrix[41];
+static float *matrix[41];
 
-static double row0[]  = {0.1437946974886250};
-static double row1[]  = {0.1513594875720590};
-static double row2[]  = {0.1593334830357050};
-static double row3[]  = {0.1677404776222450};
-static double row4[]  = {0.1766058889630630};
-static double row5[]  = {0.1859569023966960};
-static double row6[]  = {0.1958226315281160};
-static double row7[]  = {0.2062342978884670};
-static double row8[]  = {0.2172254324285080};
-static double row9[]  = {0.2288321020171980};
-static double row10[] = {0.2410931646317720};
-static double row11[] = {0.2540505575320670};
-static double row12[] = {0.2677496234274150};
-static double row13[] = {0.2822394804908960};
-static double row14[] = {0.2975734430792770};
-static double row15[] = {0.3138095012097790};
-static double row16[] = {0.3310108682663330};
-static double row17[] = {0.3492466081065300};
-static double row18[] = {0.3685923547759590};
-static double row19[] = {0.3891311404830350};
-static double row20[] = {0.4109543504366150};
-static double row21[] = {0.4341628267168120};
-static double row22[] = {0.4588681476758340};
-static double row23[] = {0.4851941146396860};
-static double row24[] = {0.5132784841210240};
-static double row25[] = {0.5432749916555080};
-static double row26[] = {0.5753557231029890};
-static double row27[] = {0.6097139012822690};
-static double row28[] = {0.6465671707381610};
-static double row29[] = {0.6861614820516400};
-static double row30[] = {0.7287757004081980};
-static double row31[] = {0.7747270924521970};
-static double row32[] = {0.8243778824987340};
-static double row33[] = {0.8781431162166330};
-static double row34[] = {0.9365001299367070};
-static double row35[] = {0.0000000000000000};
-static double row36[] = {0.0000000000000000};
-static double row37[] = {0.0000000000000000};
-static double row38[] = {0.0000000000000000};
-static double row39[] = {0.0000000000000000};
-static double row40[] = {0.0000000000000000};
+static float row0[]  = {0.1437946974886250f};
+static float row1[]  = {0.1513594875720590f};
+static float row2[]  = {0.1593334830357050f};
+static float row3[]  = {0.1677404776222450f};
+static float row4[]  = {0.1766058889630630f};
+static float row5[]  = {0.1859569023966960f};
+static float row6[]  = {0.1958226315281160f};
+static float row7[]  = {0.2062342978884670f};
+static float row8[]  = {0.2172254324285080f};
+static float row9[]  = {0.2288321020171980f};
+static float row10[] = {0.2410931646317720f};
+static float row11[] = {0.2540505575320670f};
+static float row12[] = {0.2677496234274150f};
+static float row13[] = {0.2822394804908960f};
+static float row14[] = {0.2975734430792770f};
+static float row15[] = {0.3138095012097790f};
+static float row16[] = {0.3310108682663330f};
+static float row17[] = {0.3492466081065300f};
+static float row18[] = {0.3685923547759590f};
+static float row19[] = {0.3891311404830350f};
+static float row20[] = {0.4109543504366150f};
+static float row21[] = {0.4341628267168120f};
+static float row22[] = {0.4588681476758340f};
+static float row23[] = {0.4851941146396860f};
+static float row24[] = {0.5132784841210240f};
+static float row25[] = {0.5432749916555080f};
+static float row26[] = {0.5753557231029890f};
+static float row27[] = {0.6097139012822690f};
+static float row28[] = {0.6465671707381610f};
+static float row29[] = {0.6861614820516400f};
+static float row30[] = {0.7287757004081980f};
+static float row31[] = {0.7747270924521970f};
+static float row32[] = {0.8243778824987340f};
+static float row33[] = {0.8781431162166330f};
+static float row34[] = {0.9365001299367070f};
+static float row35[] = {0.0000000000000000f};
+static float row36[] = {0.0000000000000000f};
+static float row37[] = {0.0000000000000000f};
+static float row38[] = {0.0000000000000000f};
+static float row39[] = {0.0000000000000000f};
+static float row40[] = {0.0000000000000000f};
 
 /* tv */
-double** tv_PureEndowment() {
+float** tv_PureEndowment() {
 
   matrix[0] = row0;
   matrix[1] = row1;
@@ -105,29 +105,29 @@ double** tv_PureEndowment() {
 }
 
 static int n = 35;
-static double bdeath = 1;
+static float bdeath = 1;
 
-static double b_0(double t) {
-    return 0.0;
+static float b_0(float t) {
+    return 0.0f;
 }
 
-static double mu_01(double t) {
+static float mu_01(float t) {
     return GM(t);
 }
 
-static double bj_00(double t) {
-    return t == pensiontime ? bpension: 0.0;
+static float bj_00(float t) {
+    return t == pensiontime ? bpension: 0.0f;
 }
 
-static double bj_01(double t) {
-    return 0.0; 
+static float bj_01(float t) {
+    return 0.0f; 
 }
 
-void bj_ii_PureEndowment(double t, double* result) {
+void bj_ii_PureEndowment(float t, float* result) {
   result[0] += bj_00(t);
 }
 
-void dy_PureEndowment(double t, double* V,double* result)
+void dy_PureEndowment(float t, float* V,float* result)
 {
     result[0] = r(t) * V[0] - b_0(t) - mu_01(t) * (0 - V[0] + bj_01(t));
 }
