+++ c_float/Policy_Distributor.h
@@ -3,8 +3,8 @@
 
 extern int policy;
 
-void dy(double t, double* V, double* result);
-void bj_ii(double t, double* result);
-double** test_values();
+void dy(float t, float* V, float* result);
+void bj_ii(float t, float* result);
+float** test_values();
 
 #endif
