+++ c_float/RK_Library.c
@@ -10,20 +10,20 @@
 #include "RK_Library.h"
 #include <math.h>
 
-double age = 30;
-double interestrate = 0.05;
-double bpension = 1;
-double pensiontime = 35;
+float age = 30;
+float interestrate = 0.05f;
+float bpension = 1;
+float pensiontime = 35;
 
-double GM(double t) {
-    return 0.0005 + pow(10.0, 5.728 - 10.0 + 0.038*(age + t));
+float GM(float t) {
+    return 0.0005f + pow(10.0f, 5.728f - 10.0f + 0.038f*(age + t));
 }
 
 // Interest
-double r(double t) {
+float r(float t) {
     return interestrate;
 }
 
-double indicator(int b) {
-    return b ? 1.0 : 0.0;
+float indicator(int b) {
+    return b ? 1.0f : 0.0f;
 }
